requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
expiringdict==1.2.2
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
launchdarkly-eventsource==1.2.1
launchdarkly-server-sdk==9.9.0
MarkupSafe==3.0.2
pyRFC3339==2.0.1
python-dotenv==1.0.1
requests==2.32.3
semver==3.0.4
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.21.0
